Add files via upload
Nim is a simple game that where two players are presented with a number of piles of sticks. The two players alternate removing sticks from one pile at a time according to certain rules. The goal of the game is sometimes to remove the last stick, or avoid removing the last stick. Here are the rules for our version of the game.

The game begins with three piles, labeled A, B, and C. Pile A starts with 5 sticks, B with 4, and C with 3. Player 1 moves first, and is allowed to remove any number of sticks (> 0) from a single pile. Player 2 then moves, and is allowed to make a similar choice: they choose a pile, and remove any positive number of sticks from that pile. The two players alternate turns until one player is forced to take the last stick (usually because it is the only one left). That player loses the game, and the other player wins.